Minor changes at UserRepository and UserEntity
