--- src/main/java/uz/oliymahad/userservice/model/entity/auth/AuthProviderEnum.java
@@ -1,7 +1,0 @@
-package uz.oliymahad.userservice.model.entity.auth;
-
-public enum AuthProviderEnum {
-    LOCAL,
-    GOOGLE,
-    FACEBOOK
-}
