+++ src/main/java/uz/oliymahad/userservice/model/enums/EAuthProvider.java
@@ -1,0 +1,7 @@
+package uz.oliymahad.userservice.model.enums;
+
+public enum EAuthProvider {
+    LOCAL,
+    GOOGLE,
+    FACEBOOK
+}
