+++ src/main/java/uz/oliymahad/userservice/repository/RoleRepository.java
@@ -1,7 +1,7 @@
 package uz.oliymahad.userservice.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import uz.oliymahad.userservice.model.RoleEntity;
+import uz.oliymahad.userservice.model.entity.RoleEntity;
 import uz.oliymahad.userservice.model.enums.ERole;
 
 public interface RoleRepository extends JpaRepository<RoleEntity, Integer> {
