+++ src/main/java/uz/oliymahad/userservice/repository/UserDetailRepository.java
@@ -1,7 +1,7 @@
 package uz.oliymahad.userservice.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import uz.oliymahad.userservice.model.UserDetail;
+import uz.oliymahad.userservice.model.entity.UserDetail;
 
 public interface UserDetailRepository extends JpaRepository<UserDetail, Long> {
