+++ src/main/java/uz/oliymahad/userservice/service/UserService.java
@@ -5,8 +5,8 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import uz.oliymahad.userservice.dto.UserRegisterDto;
 import uz.oliymahad.userservice.dto.response.ApiResponse;
-import uz.oliymahad.userservice.model.RoleEntity;
-import uz.oliymahad.userservice.model.UserEntity;
+import uz.oliymahad.userservice.model.entity.RoleEntity;
+import uz.oliymahad.userservice.model.entity.UserEntity;
 import uz.oliymahad.userservice.model.enums.ERole;
 import uz.oliymahad.userservice.repository.RoleRepository;
 import uz.oliymahad.userservice.repository.UserRepository;
